requirements.txt
Django==5.2.8
djangorestframework==3.15.1
psycopg2-binary==2.9.9
dj-database-url==1.3.0
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8
django-debug-toolbar==4.2.0
django-extensions==3.2.3
numpy==1.26.4
pandas==2.1.4
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1
rasterio==1.3.9
boto3==1.34.125
geopandas==0.14.3
requests==2.31.0
elevation==1.1.3
